migrate element tag data in batches
